Refactor ingredient handling to improve user association and validation; update data types for calories in IngredientRequest and IngredientResponse

+++ src/main/java/org/tcpmanager/tcpmanager/calories/ingredient/dto/IngredientResponse.java
@@ -2,7 +2,7 @@ package org.tcpmanager.tcpmanager.calories.ingredient.dto;
 
 import java.math.BigDecimal;
 
-public record IngredientResponse(Long id, String name, BigDecimal calories, BigDecimal fats,
+public record IngredientResponse(Long id, String name, Integer calories, BigDecimal fats,
                                  BigDecimal carbs, BigDecimal proteins, String ean) {
 
 }
